requirements.txt
requests==2.32.3
pandas>=2.0.0
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
db-dtypes==1.3.0
pandas_gbq
dotenv